Add ActiveIngredientsImporter
Add the ActiveIngredientsImporter class and a wrapping rake task to load
the active ingredients list and counts from the OpenFDA API using a
count query.

This query will only return the top 1000 ingredients by count as it
currently is written since the API does not allow for a limit of more
than 1000 terms for count querries.