Display results in a grid
Use bootstrap classes to display the drug results in a grid format.

Add the substance name to the pagination links so that the pagination
links work correctly even if the substance text field has been blanked
out.

Add a label to the substance textfield so we pass 508 and wcag2.0.

Move some of the helper functions that were being defined in the jQuery
onload block into FDA.Helpers so we can use them from the FDA.Labels
calls to populate the div.details block when the search is triggered
through the D3 visualization.