Fix count discrepancy between import and search
add .exact to the search parameters when fetching data from the openFDA
API.  Leaving this off was causing multi-word substance names to be
tokenized in the search.  For example, searching for "Titanium Dioxide"
would bring back all results that had "Titanium" or "Dioxide" in the
name.

Replace spaces with + per API Docs

Per the API documentation, spaces in the search terms should be replaced
with "+"