Fix tooltip positioning
Move the div for the tooltip to the bottom of the HTML content.  In
order for the tooltip to show up in the correct location, it must not
have a parent container with position relative or absolute between it
and the body tag.  Bootstrap's grid layout uses position relative, so we
have to move the tooltip div out of this inheritance chain.